projets/créer_une_332710_a71d45/requirements.txt
Flask-Bootstrap
wtforms==3.0.1
flask-bootstrap==3.3.0.0
Flask
bootstrap5==5.0.0
fontawesome==5.15.3
flask-wtf==1.0.0
Flask==2.3.3
css pour l'interface graphique
python-dotenv==1.0.0
